requirements.txt
tabulate
python-nmap
